Refactor CLI helpers and support piped spell casting
